Add simple js window confirmation to delete actions
component toggleSelect on delete button is because the click listener
exists on the whole div, so this resets the state if ‘cancel’ is
clicked in the confirmation dialog

+++ app/assets/javascripts/components/main/directives/component-card/componentCardCtrl.js
@@ -49,11 +49,13 @@
     }
 
     ctrl.delete = function () {
-      console.log('Deleting component with id: ' + ctrl.component.id);
-      $http.delete('/components/'+ ctrl.component.id).then(function(){
-        $scope.$destroy();
-      });
-      $scope.gpCourseCtrl.removeComponent(ctrl.component.id);
+      if(window.confirm("Are you sure you want to delete '" + ctrl.component.name + "'?")){
+        console.log('Deleting component with id: ' + ctrl.component.id);
+        $http.delete('/components/'+ ctrl.component.id).then(function(){
+          $scope.$destroy();
+        });
+        $scope.gpCourseCtrl.removeComponent(ctrl.component.id);
+      }
     }
 
     ctrl.save = function () {
